// src/bitacora-prohibiciones-anticipadas/dto/create-bitacora-prohibiciones-anticipada.dto.ts
export class CreateBitacoraProhibicionesAnticipadaDto {

    id_bitacora_prohibicion_anticipada: number;
        
    prohibicion_anticipada_id: number;

    motivo: string;

    detalle_motivo: string;

    datos_modificados: string;

    fecha_cambio: Date;

    organismo_id: number;

    usuario_id: number;
          
}